refactor(trip-inspection): extract dialog parts
Extract TripBasicInfo and TripInspectionTripPager from the trip inspection dialog to simplify the parent component and keep display-specific logic closer to each subcomponent.
